🚚 Rename: 파사드 경로 수정
