+++ src/modules/concert.module.ts
@@ -1,7 +1,8 @@
 import { Module } from '@nestjs/common';
-import { ConcertFacadeApp } from 'src/application/concert/concert.facade(app)';
+import { ConcertFacadeApp } from 'src/application/concert/concert.facade';
 import { ConcertService } from 'src/domain/concert/concert.service';
 import { ConcertRepositoryImpl } from 'src/infrastructure/concert/concert.repository';
+import { SeatRepositoryImpl } from 'src/infrastructure/concert/seat.repository';
 import { ConcertController } from 'src/presentation/concert/concert.constroller';
 @Module({
   imports: [],
@@ -12,6 +13,10 @@ import { ConcertController } from 'src/presentation/concert/concert.constroller'
     {
       provide: 'IConcertRepository',
       useClass: ConcertRepositoryImpl,
+    },
+    {
+      provide: 'ISeatRepository',
+      useClass: SeatRepositoryImpl,
     },
   ],
   exports: [ConcertService],
