+++ src/modules/reservation.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { ReservationFacadeApp } from 'src/application/reservation/reservation.facade(app)';
+import { ReservationFacadeApp } from 'src/application/reservation/reservation.facade';
 import { ReservationService } from 'src/domain/reservation/reservation.service';
 import { ReservationRepositoryImpl } from 'src/infrastructure/reservation/reservation.repository';
 import { ReservationController } from 'src/presentation/reservation/reservation.controller';
